src/python-flask/requirements.txt
Werkzeug
Flask~=3.0.0
flask-cors
mysql-connector
requests~=2.31.0
python-dotenv
gunicorn